refactor: Group pages
