+++ apps/web/src/studio/components/workflows/index.ts
@@ -1,1 +1,4 @@
-export * from './WorkflowsListPage';
+export * from './list/WorkflowsListPage';
+export * from './node-view/WorkflowsDetailPage';
+export * from './test-step/WorkflowsTestStepPage';
+export * from './step-editor/WorkflowsStepEditorPage';
