+++ apps/web/src/studio/components/workflows/node-view/StepNode.tsx
